Fix ulong migration fallout
- Remove redundant casts and hoist checked conversions\n- Use ulong for semantically-unsigned nonce/gas fields and adjust tests\n- Make BlockDownloader.PrepareRequest swallow provider exceptions consistently

+++ src/Nethermind/Nethermind.Facade/Simulate/SimulateBlockTracer.cs
@@ -41,7 +41,7 @@ public class SimulateBlockTracer(bool isTracingLogs) : BlockTracerBase<SimulateC
     public override void StartNewBlockTrace(Block block)
     {
         _txIndex = 0;
-        _blockNumber = (ulong)block.Number;
+        _blockNumber = block.Number;
         _blockTimestamp = block.Timestamp;
         base.StartNewBlockTrace(block);
     }
